+++ src/main.ts
@@ -1,0 +1,2 @@
+export * from './router'
+export { severities, Severity, logLevels, LogLevel, LogFn, ILogger } from './logger'